+++ Task 1/Data/Implementation/Product.cs
@@ -5,10 +5,10 @@ public class Product : IProduct
 {
     public int Id { get; set; }
     public string Name { get; set; }
-    public float Price { get; set; }
+    public double Price { get; set; }
     public uint Amount { get; set; }
 
-    public Product(int Id, string Name, float Price, uint Amount)
+    public Product(int Id, string Name, double Price, uint Amount)
     {
         this.Id = Id;
         this.Name = Name;
